+++ tests/conftest.py
@@ -1,7 +1,7 @@
 """Common fixtures to be used across tests."""
+import os
 from typing import Any, Dict
 
-import os
 import pytest
 import yaml